@@ -12,46 +12,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.9",
+    "@angular/animations": "^6.1.0",
     "@angular/cdk": "^6.4.1",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/flex-layout": "^5.0.0-beta.14",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/flex-layout": "^6.0.0-beta.17",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
     "@angular/material": "^6.4.1",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/router": "^6.1.0",
     "chartist": "^0.11.0",
-    "core-js": "^2.4.1",
+    "core-js": "^2.5.7",
     "hammerjs": "^2.0.8",
     "ng-chartist": "^2.0.0-beta.1",
-    "ng-snotify": "^4.2.0",
+    "ng-snotify": "^4.3.1",
     "ngx-avatar": "^2.9.0",
-    "rxjs": "^5.5.6",
-    "zone.js": "^0.8.19"
+    "rxjs": "^6.2.2",
+    "rxjs-compat": "^6.0.0-rc.0",
+    "rxjs-tslint": "^0.1.5",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.3",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/chartist": "^0.9.39",
-    "@types/jasmine": "~2.8.3",
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular/cli": "^6.1.1",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
+    "@types/chartist": "^0.9.42",
+    "@types/jasmine": "^2.8.8",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "@types/node": "^6.0.114",
+    "codelyzer": "^4.4.2",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "^2.0.5",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "^1.4.3",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "ts-node": "~5.0.1",
+    "tslint": "~5.11.0",
+    "typescript": "~2.7.2"
   }
 }
